Show the score in the space between rectangle pairs.
The player will see, for example, 21 22 23 on the screen. He or she is likely to concentrate on the space, so it will be easier for him or her to see the score he or she will get than if it were shown only in the corner of the screen.

After passing the 21, it becomes green. A rectangle pair generated at this point would be 24, which the player will see once the rectangles are within eyesight.